add overlay logic
