+++ go.mod
@@ -11,6 +11,7 @@ require (
 	github.com/mattn/go-sqlite3 v1.14.17
 	github.com/muesli/go-app-paths v0.2.2
 	github.com/muesli/reflow v0.3.0
+	github.com/muesli/termenv v0.15.2
 	github.com/spf13/cobra v1.7.0
 	golang.org/x/term v0.9.0
 )
@@ -26,7 +27,6 @@ require (
 	github.com/mitchellh/go-homedir v1.1.0 // indirect
 	github.com/muesli/ansi v0.0.0-20230316100256-276c6243b2f6 // indirect
 	github.com/muesli/cancelreader v0.2.2 // indirect
-	github.com/muesli/termenv v0.15.2 // indirect
 	github.com/rivo/uniseg v0.4.7 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/sync v0.3.0 // indirect
